Move wx-specific preference dialog code to ui/wx.
Also, fix some imports.
