+++ pyface/tree/tree.py
@@ -12,5 +12,5 @@
 # Import the toolkit specific version.
 from __future__ import absolute_import
 
-from .toolkit import toolkit_object
+from pyface.toolkit import toolkit_object
 Tree = toolkit_object('tree.tree:Tree')
