+++ pyface/viewer/table_viewer.py
@@ -12,5 +12,5 @@
 # Import the toolkit specific version.
 from __future__ import absolute_import
 
-from .toolkit import toolkit_object
+from pyface.toolkit import toolkit_object
 TableViewer = toolkit_object('viewer.table_viewer:TableViewer')
