+++ pyface/viewer/tree_viewer.py
@@ -12,5 +12,5 @@
 # Import the toolkit specific version.
 from __future__ import absolute_import
 
-from .toolkit import toolkit_object
+from pyface.toolkit import toolkit_object
 TreeViewer = toolkit_object('viewer.tree_viewer:TreeViewer')
